requirements.txt
bcrypt==3.1.7
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cssselect==1.1.0
cssutils==1.0.2
dnspython==1.16.0
ecdsa==0.15
emails==0.6
fastapi==0.58.1
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
idna==2.10
lxml==4.5.1
motor==2.1.0
passlib==1.7.2
premailer==3.7.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.5.1
pymongo==3.10.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-jose==3.1.0
requests==2.24.0
rsa==4.6
sentry-sdk==0.16.0
six==1.15.0
starlette==0.13.4
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
